==> 1.text_parsing.ipynb <==
### Sample 10,000 speeches from each session of congress

def sample_speechids(the_path, seat, sample_size):
    '''function to sample from file of text speeches'''
    import os
    import pandas as pd
    import numpy as np
    
    the_path = the_path
    the_dirs = os.listdir(the_path)

    the_df = pd.DataFrame()
    the_filesnames = []
    #iterate through each directory one by by
    for dir_name in the_dirs:
        if dir_name[0]=="0" or dir_name[0]=="1":
            the_filenames = the_path + "/" + dir_name
            the_filesnames.append(dir_name)

    for line in the_filesnames:
        f = pd.read_csv(filepath_or_buffer = the_path + "/" + line, sep="|", header = None, engine="python")

        df = pd.DataFrame(f)
        df = df.rename(columns=df.iloc[0]).drop(df.index[0])
        df = df[df['chamber']==seat]
        np.random.seed(9)
        sampled_df = df.sample(sample_size, replace=False)
        the_df = the_df.append(sampled_df, ignore_index=True)
    return(the_df)

samples = sample_speechids("./hein-bound", "S", 10000)

samples

samples['fullname']=samples['firstname']+' '+samples['lastname']

#change altnerative party politicans to D/R party they most closely align with
sampled_df.loc[sampled_df['fullname'] == 'BERNARD SANDERS', 'party'] = "D"
sampled_df.loc[sampled_df['fullname'] == 'JOSEPH LIEBERMAN', 'party'] = "D"
sampled_df.loc[sampled_df['fullname'] == 'HARRY BYRD', 'party'] = "D"
sampled_df.loc[sampled_df['fullname'] == 'JAMES JEFFORDS', 'party'] = "D"
sampled_df.loc[sampled_df['fullname'] == 'DEAN BARKLEYS', 'party'] = "D"
sampled_df.loc[sampled_df['fullname'] == 'JAMES BUCKLEY', 'party'] = "R"
speeches_merged.loc[speeches_merged['fullname'] == 'DEAN BARKLEY', 'party'] = "R"

#Create party label 
party=pd.get_dummies(samples['party'],prefix="party",drop_first=True)
samples[party.columns]=party

def read_speech(the_path):
    '''function to read in speech files'''
    import os
    import pandas as pd
    
    the_path = the_path
    the_dirs = os.listdir(the_path)

    the_df = pd.DataFrame()

    the_filesnames = []

    #iterate through each directory one by by
    for dir_name in the_dirs:
        if dir_name[0]=="s":
            the_filenames = the_path + "/" + dir_name
            the_filesnames.append(dir_name)

    for line in the_filesnames:
        f = open(the_path + "/" + line, "r", encoding= 'ISO-8859-1')
        speech_id = []
        speech = []
        congress = []
        for line in f:
            if line[0]=="9":
                speech_id.append(line[0:9])
                speech.append(line[10:])            
                congress.append(line[0:2])
            if line[0]=="1":
                speech_id.append(line[0:10])
                speech.append(line[11:])
                congress.append(line[0:3])

        df = pd.DataFrame(list(zip(speech_id, speech, congress)), 
                       columns =['speech_id', 'speech', 'congress'])
        df = df.drop([0]) #remove first row
        the_df = the_df.append(df, ignore_index=True)
        
    return(the_df)

speeches = read_speech("/Users/kristenakey/Documents/Documents/QMSS/NLP/Project/hein-bound")

speeches

#Merge sampled speech IDs and speeches raw
speech_merged = pd.merge(samples, speeches, on=['speech_id'], how='left')

speech_merged

#speech_merged.to_pickle(r'./pickled_data/project-180000_speech_merged.pkl')

with open("./extended_stop_words.txt", 'r') as input_file:
    ext_stop_words = input_file.read().split(',')
ext_stop_words = [str(i.lower()) for i in ext_stop_words]
ext_stop_words.append(',,')
ext_stop_words.append(',')

def speech_clean(df):
    '''function to clean/parse speech files'''
    import re
    import nltk
    from nltk.corpus import stopwords
    stop_words = nltk.corpus.stopwords.words('english')
    stop_words.extend(ext_stop_words)
    from nltk.stem import PorterStemmer
    my_stem = PorterStemmer()
    from nltk.stem import WordNetLemmatizer 
    lemmatizer = WordNetLemmatizer()
    dictionary = set(w.lower() for w in nltk.corpus.words.words())
    from tqdm import tqdm

    the_df = pd.DataFrame()
    speechl = pd.Series.tolist(df[['speech']])
    
    #for check
    congressl = pd.Series.tolist(df[['congress']]) # as check
    congressl = [item for sublist in congressl for item in sublist]

    #for check
    speech_idl = pd.Series.tolist(df[['speech_id']]) 
    speech_idl = [item for sublist in speech_idl for item in sublist]

    for line, idnum, session in tqdm(zip(speechl, speech_idl, congressl), total=170000, position=0, leave=True):
        speech_og = []
        for i in line:
            if type(i)==float:
                speech_og.append("NA")
            else:
                i = i[:-1]
                speech_og.append(i)

        #Clean text
        tmp_read =re.sub("[\.]", ' ', str(line)).lower()
        tmp_read = re.sub("['\n']", '', tmp_read)
        tmp_read = re.sub('[^a-zA-Z]+', ' ', tmp_read)

        #Tokenization and remove stop words
        tmp_read = [word for word in tmp_read.split() if word not in stop_words]

            #dictionary words
        dict_read = [word for word in tmp_read if word in dictionary]

        #Stemming
        tmp_read_stm = [my_stem.stem(word) for word in tmp_read]
        dict_read_stm = [my_stem.stem(word) for word in dict_read]

        #Lemminization 
        tmp_read_lem = [lemmatizer.lemmatize(word) for word in tmp_read]
        dict_read_lem = [lemmatizer.lemmatize(word) for word in dict_read]

        #rejoin speeches
        tmp_read = ' '.join(tmp_read)
        tmp_read_stm = ' '.join(tmp_read_stm)
        tmp_read_lem = ' '.join(tmp_read_lem)

        dict_read = ' '.join(dict_read)
        dict_read_stm = ' '.join(dict_read_stm)
        dict_read_lem = ' '.join(dict_read_lem)

        #df
        tmp = pd.DataFrame([speech_og], columns=['original speech'])
        tmp['body'] = tmp_read
        tmp['body_stem'] = tmp_read_stm
        tmp['body_lem'] = tmp_read_lem
        tmp['body_dict'] = dict_read
        tmp['body_dict_stem'] = dict_read_stm
        tmp['body_dict_lem'] = dict_read_lem
        
        #for check
        tmp['congress'] = session
        tmp['speech_id'] = idnum

        the_df = the_df.append(tmp, ignore_index=True)
    return(the_df)
    #the_df.to_csv(r'/Users/kristenakey/Desktop/hein-bound-111.csv', index=False)

speech_parsed = speech_clean(speech_merged)

speech_parsed

#speech_parsed.to_pickle(r'./pickled_data/project-180000_speeches_parsed.pkl')

#rejoin clean/parsed speeches to speech_merged 
speech_c = pd.merge(speech_merged, speech_parsed, left_on=['speech_id', "congress"], right_on=['speech_id', "congress"], how='left')#[['cfscore','seat','Cand.ID','Name']]

#speech_c.to_pickle(r'./pickled_data/project-180000_speech_c.pkl')

### Split speeches from each session of congress into unique dataframes 

speech_c = pd.read_pickle(r'./pickled_data/project-180000_speech_c.pkl')

i.e df_103 = 103rd congressional speeches
