==> 2.ideology_matching.ipynb <==
import pandas as pd

#load ideology data
cf_df = pd.read_csv("https://dataverse.harvard.edu/api/access/datafile/:persistentId?persistentId=doi:10.7910/DVN/UQLKRY/MGQCGW")

def cf_filter(df, seat):
    cf_cong = cf_df[['Name','Cand.ID','cfscore', 'seat', 'Party']]
    cf_cong = cf_cong[cf_cong['seat']=="federal:senate"]
    cf_cong = cf_cong.drop_duplicates() 
    return(cf_cong)

cf_cong = cf_filter(cf_df, "federal:senate")

import re

name_correction = pd.DataFrame()
names = []
firstname = []
lastname = []
middlename = []
for i in cf_cong['Name']:
    names.append(i)
    newi = re.sub(",", " ", i)
    newi = newi.split()
    firstname.append(newi[1])
    lastname.append(newi[0])
    if len(newi) >= 3:
        middlename.append(newi[3:4])
    else:
        middlename.append("")

    tmp = pd.DataFrame(lastname, columns=['lastname'])
    tmp['firstname'] = firstname
    #tmp['middlename'] = middlename
    tmp['Name']= names

tmp['fullname'] = tmp['firstname'] +' '+tmp['lastname']

#Fix name miscodings in tmp file so they match

tmp.loc[tmp['Name']=="THURMOND, JAMES STROM", 'fullname'] = "J. THURMOND"
tmp.loc[tmp['Name']=="SYMMS, STEVE", 'fullname'] = "STEVEN SYMMS"
tmp.loc[tmp['Name']=="EXON, JAMES J.", 'fullname'] = "J. EXON"
tmp.loc[tmp['Name']=="JOHNSTON, J. BENNETT JR.", 'fullname'] = "JOHN JOHNSTON"
tmp.loc[tmp['Name']=="HECHT, CHIC", 'fullname'] = "JACOB HECHT"
tmp.loc[tmp['Name']=="GRAMM, WILLIAM PHILLIP", 'fullname'] = "PHIL GRAMM"
tmp.loc[tmp['Name']=="D'AMATO, ALFONSE M", 'fullname'] = "ALFONSE DAMATO"
tmp.loc[tmp['Name']=="NICKLES, DONALD LEE", 'fullname'] = "DON NICKLES"
tmp.loc[tmp['Name']=="GARN, E. J. JAKE", 'fullname'] = "EDWIN GARN"
tmp.loc[tmp['Name']=="ADAMS, BROCK", 'fullname'] = "BROCKMAN ADAMS"
tmp.loc[tmp['Name']=="BRADLEY, BILL", 'fullname'] = "WILLIAM BRADLEY"
tmp.loc[tmp['Name']=="NUNN, SAM A. JR.", 'fullname'] = "SAMUEL NUNN"
tmp.loc[tmp['Name']=="QUAYLE, DAN (VICE PRES)", 'fullname'] = "JAMES QUAYLE"
tmp.loc[tmp['Name']=="GORTON, SLADE", 'fullname'] = "T. GORTON"
tmp.loc[tmp['Name']=="KERREY, J ROBERT (BOB) SENATOR", 'fullname'] = "J. KERREY"
tmp.loc[tmp['Name']=="BROWN, HANK", 'fullname'] = "GEORGE BROWN"
tmp.loc[tmp['Name']=="BURDICK, QUENTIN N.", 'fullname'] = "QUENTIN BURDICK" 
# tmp.loc[tmp['Name']=="BURDICK, QUENTIN N.", 'fullname'] = "JOCELYN BURDICK" #wife took over after his death #NEED TO MANUALLY ADD AFTER JOIN
tmp.loc[tmp['Name']=="BRAUN, CAROL MOSELEY", 'fullname'] = "CAROL MOSELEY BRAUN" 
tmp.loc[tmp['Name']=="GRAMS, RODNEY", 'fullname'] = "ROD GRAMS"
tmp.loc[tmp['Name']=="DEWINE, MIKE", 'fullname'] = "MICHAEL DEWINE"
tmp.loc[tmp['Name']=="ABRAHAM, EDMOND SPENCER", 'fullname'] = "SPENCER ABRAHAM"
tmp.loc[tmp['Name']=="HUTCHINSON, Y TIM", 'fullname'] = "TIM HUTCHINSON"
tmp.loc[tmp['Name']=="ALLARD, A WAYNE", 'fullname'] = "WAYNE ALLARD"
tmp.loc[tmp['Name']=="CLELAND, JOSEPH MAXWELL (AKA MAX)", 'fullname'] = "J. CLELAND"
tmp.loc[tmp['Name']=="WILLIAMS, SEN. HARRISON A. JR.", 'fullname'] = "HARRISON WILLIAMS"
tmp.loc[tmp['Name']=="HATHAWAY, W JOHN", 'fullname'] = "WILLIAM HATHAWAY"
tmp.loc[tmp['Name']=="GRAVEL, MIKE", 'fullname'] = "MAURICE GRAVEL"
tmp.loc[tmp['Name']=="EASTLAND, HIRAM", 'fullname'] = "JAMES EASTLAND"
tmp.loc[tmp['Name']=="HAYAKAWA, S. I.", 'fullname'] = "SAMUEL HAYAKAWA"
# tmp.loc[tmp['Name']=="HUMPHREY, HUBERT H III", 'fullname'] = "MURIEL HUMPHREY" #wife took over after his death #NEED TO MANUALLY ADD AFTER JOIN

##111th Congress 
tmp.loc[tmp['Name']=="REED, JACK", 'fullname'] = "JOHN REED"
tmp.loc[tmp['Name']=="KOHL, HERB", 'fullname'] = "HERBERT KOHL"
tmp.loc[tmp['Name']=="CONRAD, GAYLORD KENT", 'fullname'] = "KENT CONRAD"
tmp.loc[tmp['Name']=="SESSIONS, JEFF", 'fullname'] = "JEFFERSON SESSIONS"
tmp.loc[tmp['Name']=="DEMINT, JAMES W", 'fullname'] = "JIM DEMINT"
tmp.loc[tmp['Name']=="CORKER, ROBERT P JR", 'fullname'] = "BOB CORKER"
tmp.loc[tmp['Name']=="WEBB, JAMES H JR", 'fullname'] = "JIM WEBB"
tmp.loc[tmp['Name']=="KAUFMAN, JOE", 'fullname'] = "EDWARD KAUFMAN"
tmp.loc[tmp['Name']=="COCHRAN, WILLIAM THAD", 'fullname'] = "THAD COCHRAN"
tmp.loc[tmp['Name']=="NELSON, E BENJAMIN", 'fullname'] = "BEN NELSON"
tmp.loc[tmp['Name']=="JOHANNS, MICHAEL O", 'fullname'] = "MIKE JOHANNS"
tmp.loc[tmp['Name']=="MERKLEY, JEFFREY ALAN", 'fullname'] = "JEFF MERKLEY"
tmp.loc[tmp['fullname']=="CONRAD, GAYLORD KENT", 'fullname'] = "KENT CONRAD"
tmp.loc[tmp['Name']=="FRANKEN, AL", 'fullname'] = "ALAN FRANKEN"

cf_cong = pd.merge(cf_cong,tmp, on=['Name'], how='outer')

#### Load in parsed speech data (completed with text_parsing script)

speeches_c = pd.read_pickle(r'/Users/kristenakey/Documents/Documents/QMSS/NLP/Project/pickled_data/project-180000_speech_c.pkl')

speeches_c = speeches_c[speeches_c.congress != '94']

speeches_c['fullname'] = speeches_c['firstname'] +' '+ speeches_c['lastname']

#join by fullname or last name --- first and middle not always the same (i.e Jeff vs Jeffry Flake)
sp_ideo = pd.merge(speeches_c, cf_cong, how='left', on=['fullname'])#[['cfscore','seat','Cand.ID','Name']]

#### Names without cfscores

#no CFscore available 
sp_ideo.loc[sp_ideo['fullname']=="WILLIAM PROXMIRE", 'cfscore'] = 0
sp_ideo.loc[sp_ideo['fullname']=="JOCELYN BURDICK", 'cfscore'] = 0
sp_ideo.loc[sp_ideo['fullname']=="EDWARD KAUFMAN", 'cfscore'] = 0
sp_ideo.loc[sp_ideo['fullname']=="CARTE GOODWIN", 'cfscore'] = 0
sp_ideo.loc[sp_ideo['fullname']=="ADLAI STEVENSON", 'cfscore'] = 0
sp_ideo.loc[sp_ideo['fullname']=="JAMES ALLEN", 'cfscore'] = 0
sp_ideo.loc[sp_ideo['fullname']=="THOMAS MCINTYRE", 'cfscore'] = 0
sp_ideo.loc[sp_ideo['fullname']=="JOHN SPARKMAN", 'cfscore'] = 0
sp_ideo.loc[sp_ideo['fullname']=="HARRISON WILLIAMS", 'cfscore'] = 0
sp_ideo.loc[sp_ideo['fullname']=="HARRY BYRD", 'cfscore'] = 0
sp_ideo.loc[sp_ideo['fullname']=="RICHARD CLARK", 'cfscore'] = 0
sp_ideo.loc[sp_ideo['fullname']=="HENRY BELLMON", 'cfscore'] = 0
sp_ideo.loc[sp_ideo['fullname']=="CLIFFORD CASE", 'cfscore'] = 0
sp_ideo.loc[sp_ideo['fullname']=="CARL CURTIS", 'cfscore'] = 0
sp_ideo.loc[sp_ideo['fullname']=="JAMES ABOUREZK", 'cfscore'] = 0
sp_ideo.loc[sp_ideo['fullname']=="ABRAHAM RIBICOFF", 'cfscore'] = 0
sp_ideo.loc[sp_ideo['fullname']=="FLOYD HASKELL", 'cfscore'] = 0
sp_ideo.loc[sp_ideo['fullname']=="LEE METCALF", 'cfscore'] = 0
sp_ideo.loc[sp_ideo['fullname']=="DEWEY BARTLETT", 'cfscore'] = 0
sp_ideo.loc[sp_ideo['fullname']=="LEE METCALF", 'cfscore'] = 0
sp_ideo.loc[sp_ideo['fullname']=="CLIFFORD HANSEN", 'cfscore'] = 0
sp_ideo.loc[sp_ideo['fullname']=="JOHN MCCLELLAN", 'cfscore'] = 0
sp_ideo.loc[sp_ideo['fullname']=="MILTON YOUNG", 'cfscore'] = 0
sp_ideo.loc[sp_ideo['fullname']=="RICHARD SCHWEIKER", 'cfscore'] = 0
sp_ideo.loc[sp_ideo['fullname']=="JAMES PEARSON", 'cfscore'] = 0
sp_ideo.loc[sp_ideo['fullname']=="KANEASTER HODGES", 'cfscore'] = 0
sp_ideo.loc[sp_ideo['fullname']=="MURIEL HUMPHREY", 'cfscore'] = 0
sp_ideo.loc[sp_ideo['fullname']=="PAUL HATFIELD", 'cfscore'] = 0
sp_ideo.loc[sp_ideo['fullname']=="NICHOLAS BRADY", 'cfscore'] = 0
sp_ideo.loc[sp_ideo['fullname']=="EDMUND MUSKIE", 'cfscore'] = 0

nan_cfscore = []
for i in sp_ideo[sp_ideo['cfscore']==0]['fullname'].unique():
    nan_cfscore.append(i)

**Check all politicans have CFscore**

sp_ideo = pd.DataFrame.drop(sp_ideo, columns=["district", "Party"])

sp_ideo = sp_ideo.drop_duplicates(['speech_id']).sort_values(by=['speech_id'])

sp_ideo[sp_ideo.isnull().any(axis=1)].fullname.isna().sum()

sp_ideo['cfscore'].isna().sum()

sp_ideo[sp_ideo['cfscore'].isna()].fullname.unique()

## Categorize Ideology

def ideo_vec(df):
    '''Categorizes ideology into three categories:
    extreme liberal, moderate, extreme conservative.
    Two measurements: ideo_vec, ideo_vec_ext'''
    newdf = pd.DataFrame()
    for session, df in df.groupby('congress'):
        print(session)
        grouped_speeches = df.groupby(['party_R']) #to calculate median cfscores by party
        l = pd.Series.tolist(df[['cfscore','party_R','fullname','speechid']])
        ideo_vec = []
        ideo_vec_ext = []
        count = 0
        for i in l:

            if i[2] not in nan_cfscore:
            
            #ideo_vec
                #conservatives - ext.
                if (i[0] >= grouped_speeches['cfscore'].quantile([.25, .5, .75])[1].iloc[1]):
                    ideo_vec.append(1)

                #mod.
                if (i[0] < grouped_speeches['cfscore'].quantile([.25, .5, .75])[1].iloc[1]) and (i[0]> grouped_speeches['cfscore'].quantile([.25, .5, .75])[0].iloc[1]):
                    ideo_vec.append(0) 

                #liberal - ext.
                if (i[0] <= grouped_speeches['cfscore'].quantile([.25, .5, .75])[0].iloc[1]):
                    ideo_vec.append(-1)

            #ideo_vec_ext
                #conservatives - ext.
                if (i[0] >= grouped_speeches['cfscore'].quantile([.25, .5, .75])[1].iloc[2]):
                    ideo_vec_ext.append(1)

                #mod.
                if (i[0] < grouped_speeches['cfscore'].quantile([.25, .5, .75])[1].iloc[0]) and (i[0]> grouped_speeches['cfscore'].quantile([.25, .5, .75])[0].iloc[2]):
                    ideo_vec_ext.append(0)

                #liberal - ext.
                if (i[0] <= grouped_speeches['cfscore'].quantile([.25, .5, .75])[0].iloc[0]):
                    ideo_vec_ext.append(-1)
                    
                    
                if (i[0] <= grouped_speeches['cfscore'].quantile([.25, .5, .75])[0].iloc[2]) and (i[0]>grouped_speeches['cfscore'].quantile([.25, .5, .75])[0].iloc[0]):
                    ideo_vec_ext.append("NA")

                if (i[0] >= grouped_speeches['cfscore'].quantile([.25, .5, .75])[1].iloc[0]) and (i[0]<grouped_speeches['cfscore'].quantile([.25, .5, .75])[1].iloc[2]):
                    ideo_vec_ext.append("NA")

            if i[2] in nan_cfscore:
                ideo_vec.append("NA")
                ideo_vec_ext.append("NA")

            count += 1
            if count % 2000==0:
                print(count)

        df['ideo_vec']= ideo_vec
        df['ideo_vec_ext']= ideo_vec_ext
        newdf = newdf.append(df, ignore_index=True)
        print('df done')
        
    return(newdf)

df_final = ideo_vec(sp_ideo)

df_final.to_pickle(r'./pickled_data/project-170000_final-df.pkl')

df_final

### Add Summary Statistic Variables

def speech_length(df_c):
    '''count number of words in each speech'''
    l = pd.Series.tolist(df_c[['original speech']])
    ct = 0
    count_per_speech = []
    for line in l:
        for i in line:
            i = i.split()
            ct += 1
            count_per_speech.append(len(i))
    df_c['speech_length'] = count_per_speech
    return(df_c)

def summary_stats(df_c):
    '''sent_length = num of sentences in speech
       num_word_unique = num of unique words in speech'''
    from nltk.tokenize import sent_tokenize, word_tokenize
    
    sent_len = []
    unique_len = []
    body_basic = []
    punctuations = '''!()-[]{}';:'"\,<>./?@#$%^&*_~``'''
    ct = 0
    for i in df_c['original speech']:
        sents = sent_tokenize(i)
        tmp = word_tokenize(i)
        tmp_words = [i.lower() for i in tmp if i not in punctuations]
        tmp_unique = len(set(tmp))
        tmp_words = ' '.join(tmp_words)
        sent_len.append(len(sents))
        unique_len.append(tmp_unique)
        body_basic.append(tmp_words)
        
        ct += 1
        if ct % 1000==0:
            print(ct)
            
    df_c['body_basic'] = body_basic
    df_c['sent_length'] = sent_len
    df_c['num_word_unique'] = unique_len
    return(df_c)

df_final = speech_length(df_final)
df_final = summary_stats(df_final)

df_final['congress'] = df_final['congress'].astype(int)
df_final = df_final.sort_values(by="congress")

### Split speeches from each session of congress into unique dataframes 

for i, g in df_final.groupby('congress'):
    globals()['df_' + str(i)] =  g

##save all dfs to one pickle: "all_dfs.dat"

import pickle
PIK = "pickled_data/all_dfs.dat"

data = [df_final, df_95, df_96, df_97, df_98, df_99, df_101,
       df_101, df_102 ,  df_103 ,  df_104 ,  df_105 ,  df_106 ,  df_107 ,  df_108 , 
        df_109 ,  df_110 ,  df_111]
with open(PIK, "wb") as f:
    pickle.dump(data, f)
